created method to initialize db
This method creates the necessary tables that are used in the
project. Since these tables currently are not the way the tables
normally were (due to the change in overall styles), the methods
cannot be used yet. Some todo's are left for reference to what to
do next.

Essentially, we now have only 5 tables:
Courses,
Questions,
AnswerChoices,
Responses,
ResponseMappings